requirements.server.txt
# These requirements are what you would install on a production server.
# The server has more requirements than the client.
.
-r requirements.client.txt
gunicorn==23.0.0
psycopg2==2.9.10
flask-compress==1.17
